Crucible/All: Deduplication of override matching code
