+++ src/SAWScript/Crucible/LLVM/Override.hs
@@ -97,37 +97,21 @@ import           Data.Parameterized.Classes ((:~:)(..), testEquality)
 import qualified Data.Parameterized.Context as Ctx
 import           Data.Parameterized.Some (Some(..))
 
-import           Verifier.SAW.Prelude (scEq)
 import           Verifier.SAW.SharedTerm
-import           Verifier.SAW.TypedAST
 import           Verifier.SAW.Recognizer
 import           Verifier.SAW.TypedTerm
 
-import           SAWScript.Crucible.Common (Sym)
+import           SAWScript.Crucible.Common (Sym, LabeledPred, LabeledPred', labelWithSimError)
 import           SAWScript.Crucible.Common.MethodSpec (SetupValue(..), PointsTo)
 import qualified SAWScript.Crucible.Common.MethodSpec as MS
 import           SAWScript.Crucible.Common.MethodSpec (AllocIndex(..), PrePost(..))
-import           SAWScript.Crucible.Common.Override hiding (getSymInterface)
-import qualified SAWScript.Crucible.Common.Override as Ov (getSymInterface)
+import           SAWScript.Crucible.Common.Override
 import           SAWScript.Crucible.LLVM.MethodSpecIR
 import           SAWScript.Crucible.LLVM.ResolveSetupValue
 import           SAWScript.Options
 import           SAWScript.Utils (bullets, handleException)
 
 ------------------------------------------------------------------------
-
-type LabeledPred' sym = W4.LabeledPred (W4.Pred sym) PP.Doc
-type LabeledPred sym = W4.LabeledPred (W4.Pred sym) Crucible.SimError
-
--- | Convert a predicate with a 'PP.Doc' label to one with a 'Crucible.SimError'
-labelWithSimError ::
-  W4.ProgramLoc ->
-  (PP.Doc -> String) ->
-  LabeledPred' Sym ->
-  LabeledPred Sym
-labelWithSimError loc conv lp =
-  lp & W4.labeledPredMsg
-     %~ (Crucible.SimError loc . Crucible.AssertFailureSimError . conv)
 
 -- | Retrieve pre-state assertions, ready to be passed to 'Crucible.addAssertion'
 prestateAsserts :: OverrideState (LLVM arch) -> [LabeledPred Sym]
@@ -163,7 +147,7 @@ ppLLVMVal ::
   LLVMVal ->
   OverrideMatcher (LLVM arch) w PP.Doc
 ppLLVMVal cc val = do
-  sym <- Ov.getSymInterface
+  sym <- getSymInterface
   mem <- readGlobal (Crucible.llvmMemVar (cc^.ccLLVMContext))
   liftIO $ Crucible.ppLLVMValWithGlobals sym (Crucible.memImplGlobalMap mem) val
 
@@ -564,7 +548,7 @@ methodSpecHandler_prestate ::
 methodSpecHandler_prestate opts sc cc args cs =
     do let expectedArgTypes = Map.elems (cs ^. MS.csArgBindings)
 
-       sym <- Ov.getSymInterface
+       sym <- getSymInterface
 
        let aux (memTy, setupVal) (Crucible.AnyValue tyrep val) =
              do storTy <- Crucible.toStorableType memTy
@@ -698,7 +682,7 @@ enforceDisjointness ::
   MS.StateSpec (LLVM arch) ->
   OverrideMatcher (LLVM arch) md ()
 enforceDisjointness loc ss =
-  do sym <- Ov.getSymInterface
+  do sym <- getSymInterface
      sub <- OM (use setupValueSub)
      let (allocsRW, allocsRO) = Map.partition (view isMut) (view MS.csAllocs ss)
          memsRW = Map.elems $ Map.intersectionWith (,) allocsRW sub
@@ -844,27 +828,6 @@ assignVar cc loc var val =
             , "  " ++ show (Crucible.ppPtr val')
             ]
 
-------------------------------------------------------------------------
-
-
-assignTerm ::
-  SharedContext      {- ^ context for constructing SAW terms    -} ->
-  LLVMCrucibleContext arch   {- ^ context for interacting with Crucible -} ->
-  PrePost                                                          ->
-  VarIndex {- ^ external constant index -} ->
-  Term     {- ^ value                   -} ->
-  OverrideMatcher (LLVM arch) md (Maybe (LabeledPred' Sym))
-
-assignTerm sc cc prepost var val =
-  do mb <- OM (use (termSub . at var))
-     case mb of
-       Nothing -> OM (termSub . at var ?= val) >> pure Nothing
-       Just old -> matchTerm sc cc prepost val old
-
---          do t <- liftIO $ scEq sc old val
---             p <- liftIO $ resolveSAWPred cc t
---             addAssert p (Crucible.AssertFailureSimError ("literal equality " ++ stateCond prepost))
-
 
 ------------------------------------------------------------------------
 
@@ -888,11 +851,11 @@ matchArg opts sc cc cs prepost actual expectedTy expected =
     (_, _, SetupTerm expectedTT)
       | Cryptol.Forall [] [] tyexpr <- ttSchema expectedTT
       , Right tval <- Cryptol.evalType mempty tyexpr
-        -> do sym      <- Ov.getSymInterface
+        -> do sym      <- getSymInterface
               failMsg  <- mkStructuralMismatch opts cc sc cs actual expected expectedTy
               realTerm <- valueToSC sym (cs ^. MS.csLoc) failMsg tval actual
               maybeToList <$>
-                matchTerm sc cc prepost realTerm (ttTerm expectedTT)
+                matchTerm sc (cc^.ccBackend) prepost realTerm (ttTerm expectedTT)
 
     -- match the fields of struct point-wise
     (Crucible.LLVMValStruct xs, Crucible.StructType fields, SetupStruct () _ zs) ->
@@ -914,7 +877,7 @@ matchArg opts sc cc cs prepost actual expectedTy expected =
              return Nothing
 
         SetupNull () | Just Refl <- testEquality (W4.bvWidth off) Crucible.PtrWidth ->
-          do sym <- Ov.getSymInterface
+          do sym <- getSymInterface
              lp <- W4.LabeledPred
                <$> liftIO (Crucible.ptrIsNull sym Crucible.PtrWidth (Crucible.LLVMPointer blk off))
                <*> notEqual prepost opts cc sc cs expected actual
@@ -922,7 +885,7 @@ matchArg opts sc cc cs prepost actual expectedTy expected =
 
         SetupGlobal () name | Just Refl <- testEquality (W4.bvWidth off) Crucible.PtrWidth ->
           do let mem = cc^.ccLLVMEmptyMem
-             sym  <- Ov.getSymInterface
+             sym  <- getSymInterface
              ptr2 <- liftIO $ Crucible.doResolveGlobal sym mem (L.Symbol name)
              lp <- W4.LabeledPred
                <$> liftIO (Crucible.ptrEq sym Crucible.PtrWidth (Crucible.LLVMPointer blk off) ptr2)
@@ -938,7 +901,7 @@ matchArg opts sc cc cs prepost actual expectedTy expected =
   where
     resolveAndMatch = do
       (ty, val) <- resolveSetupValueLLVM opts cc sc cs expected
-      sym  <- Ov.getSymInterface
+      sym  <- getSymInterface
       if expectedTy /= ty
       then failure (cs ^. MS.csLoc) =<<
             mkStructuralMismatch opts cc sc cs actual expected expectedTy
@@ -1033,37 +996,6 @@ typeToSC sc t =
 
 ------------------------------------------------------------------------
 
-matchTerm ::
-  SharedContext   {- ^ context for constructing SAW terms    -} ->
-  LLVMCrucibleContext arch {- ^ context for interacting with Crucible -} ->
-  PrePost                                                       ->
-  Term            {- ^ exported concrete term                -} ->
-  Term            {- ^ expected specification term           -} ->
-  OverrideMatcher (LLVM arch) md (Maybe (LabeledPred' Sym))
-
-matchTerm _ _ _ real expect | real == expect = return Nothing
-matchTerm sc cc prepost real expect =
-  do free <- OM (use osFree)
-     case unwrapTermF expect of
-       FTermF (ExtCns ec)
-         | Set.member (ecVarIndex ec) free ->
-         do assignTerm sc cc prepost (ecVarIndex ec) real
-
-       _ ->
-         do t <- liftIO $ scEq sc real expect
-            p <- liftIO $ resolveSAWPred cc t
-            return $ Just $ W4.LabeledPred p $ PP.vcat $ map PP.text
-              [ "Literal equality " ++ stateCond prepost
-              , "Expected term: " ++ prettyTerm expect
-              , "Actual term:   " ++ prettyTerm real
-              ]
-  where prettyTerm term =
-          let pretty_ = show (ppTerm defaultPPOpts term)
-          in if length pretty_ < 200 then pretty_ else "<term omitted due to size>"
-
-
-------------------------------------------------------------------------
-
 -- | Use the current state to learn about variable assignments based on
 -- preconditions for a procedure specification.
 learnSetupCondition ::
@@ -1093,7 +1025,7 @@ learnGhost ::
   OverrideMatcher (LLVM arch) md (Maybe (LabeledPred Sym))
 learnGhost sc cc loc prepost var expected =
   do actual <- readGlobal var
-     maybeAssert <-  matchTerm sc cc prepost (ttTerm actual) (ttTerm expected)
+     maybeAssert <-  matchTerm sc (cc^.ccBackend) prepost (ttTerm actual) (ttTerm expected)
      pure $ fmap (labelWithSimError loc show) maybeAssert
 
 ------------------------------------------------------------------------
@@ -1122,7 +1054,7 @@ learnPointsTo opts sc cc spec prepost (LLVMPointsTo loc ptr val) =
      -- In case the types are different (from crucible_points_to_untyped)
      -- then the load type should be determined by the rhs.
      storTy <- Crucible.toStorableType memTy
-     sym    <- Ov.getSymInterface
+     sym    <- getSymInterface
 
      mem    <- readGlobal $ Crucible.llvmMemVar
                           $ (cc^.ccLLVMContext)
@@ -1175,10 +1107,6 @@ learnPointsTo opts sc cc spec prepost (LLVMPointsTo loc ptr val) =
 
 ------------------------------------------------------------------------
 
-stateCond :: PrePost -> String
-stateCond PreState = "precondition"
-stateCond PostState = "postcondition"
-
 -- | Process a "crucible_equal" statement from the precondition
 -- section of the CrucibleSetup block.
 learnEqual ::
@@ -1212,7 +1140,7 @@ learnPred ::
 learnPred sc cc loc prepost t =
   do s <- OM (use termSub)
      u <- liftIO $ scInstantiateExt sc s t
-     p <- liftIO $ resolveSAWPred cc u
+     p <- liftIO $ resolveSAWPred (cc ^. ccBackend) u
      return $
        W4.LabeledPred p (Crucible.SimError loc (Crucible.AssertFailureSimError (stateCond prepost)))
 
@@ -1378,7 +1306,7 @@ executePred ::
 executePred sc cc tt =
   do s <- OM (use termSub)
      t <- liftIO $ scInstantiateExt sc s (ttTerm tt)
-     p <- liftIO $ resolveSAWPred cc t
+     p <- liftIO $ resolveSAWPred (cc ^. ccBackend) t
      addAssume p
 
 ------------------------------------------------------------------------
@@ -1396,29 +1324,6 @@ executeFreshPointer cc (AllocIndex i) =
      blk <- W4.freshConstant sym (mkName "blk") W4.BaseNatRepr
      off <- W4.freshConstant sym (mkName "off") (W4.BaseBVRepr Crucible.PtrWidth)
      return (Crucible.LLVMPointer blk off)
-
-------------------------------------------------------------------------
-
--- | Map the given substitution over all 'SetupTerm' constructors in
--- the given 'SetupValue'.
-instantiateSetupValue ::
-  SharedContext     ->
-  Map VarIndex Term ->
-  SetupValue (LLVM arch)        ->
-  IO (SetupValue (LLVM arch))
-instantiateSetupValue sc s v =
-  case v of
-    SetupVar{}               -> return v
-    SetupTerm tt             -> SetupTerm        <$> doTerm tt
-    SetupStruct () p vs      -> SetupStruct () p <$> mapM (instantiateSetupValue sc s) vs
-    SetupArray () vs         -> SetupArray ()    <$> mapM (instantiateSetupValue sc s) vs
-    SetupElem{}              -> return v
-    SetupField{}             -> return v
-    SetupNull{}              -> return v
-    SetupGlobal{}            -> return v
-    SetupGlobalInitializer{} -> return v
-  where
-    doTerm (TypedTerm schema t) = TypedTerm schema <$> scInstantiateExt sc s t
 
 ------------------------------------------------------------------------
 
@@ -1451,7 +1356,7 @@ resolveSetupValue ::
   OverrideMatcher (LLVM arch) md (Crucible.MemType, Crucible.RegValue Sym tp)
 resolveSetupValue opts cc sc spec tp sval =
   do (memTy, lval) <- resolveSetupValueLLVM opts cc sc spec sval
-     sym <- Ov.getSymInterface
+     sym <- getSymInterface
      val <- liftIO $ Crucible.unpackMemValue sym tp lval
      return (memTy, val)
